move annotate_ast to vyper/ast/annotation.py
